tea_leaves_diseases/deployment/frontend_p2m2/requirements.txt
requests
streamlit
typing_extensions
numpy
tensorflow-cpu
pillow